fix: 메소드에 작품UUID 추가, 삭제된 그룹과 캐릭터 예외 처리, uuid 프론트에서 받아쓰기

+++ server/src/main/java/com/galaxy/novelit/character/service/CharacterServiceImpl.java
@@ -18,6 +18,8 @@ import com.galaxy.novelit.character.entity.RelationEntity.Relation;
 import com.galaxy.novelit.character.repository.CharacterRepository;
 import com.galaxy.novelit.character.repository.GroupRepository;
 import com.galaxy.novelit.character.repository.RelationRepository;
+import com.galaxy.novelit.common.exception.DeletedElementException;
+import com.galaxy.novelit.common.exception.NoSuchElementFoundException;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
@@ -41,32 +43,30 @@ public class CharacterServiceImpl implements CharacterService {
 
     @Transactional(readOnly = true)
     @Override
-    public CharacterDtoRes getCharacterInfo(String characterUUID, String userUUID) {
-        CharacterEntity character = characterRepository.findByCharacterUUID(characterUUID);
+    public CharacterDtoRes getCharacterInfo(String characterUUID, String userUUID, String workspaceUUID) {
+        CharacterEntity character = characterRepository.findByCharacterUUIDAndWorkspaceUUID(characterUUID, workspaceUUID);
 
-        CharacterDtoRes dto = new CharacterDtoRes();
-        dto.setGroupUUID(character.getGroupUUID());
-        dto.setCharacterName(character.getCharacterName());
-        dto.setCharacterUUID(character.getCharacterUUID());
-        dto.setInformation(character.getInformation());
-        dto.setDescription(character.getDescription());
-        dto.setRelations(character.getRelationship().getRelations());
-        dto.setDeleted(character.isDeleted());
-        dto.setCharacterImage(character.getCharacterImage());
+        checkCharacterException(character);
 
-//        Map<String, Double> xy = new HashMap<>();
-//        xy.put("x", character.getCharacterNode().get("x"));
-//        xy.put("y", character.getCharacterNode().get("y"));
-//        dto.setCharacterNode(xy);
-//        dto.setCharacterNode(character.getCharacterNode());
-        return dto;
+        return CharacterDtoRes.builder()
+            .workspaceUUID(character.getWorkspaceUUID())
+            .groupUUID(character.getGroupUUID())
+            .characterUUID(character.getCharacterUUID())
+            .characterName(character.getCharacterName())
+            .description(character.getDescription())
+            .information(character.getInformation())
+            .relations(character.getRelationship().getRelations())
+            .characterImage(character.getCharacterImage())
+            .build();
     }
 
     @Transactional(readOnly = true)
     @Override
-    public List<CharacterThumbnailDtoRes> getCharacters(String groupUUID, String userUUID) {
-        List<CharacterEntity> characters = characterRepository.findAllByGroupUUID(groupUUID);
-//            .orElseThrow(() -> new NoSuchElementFoundException("없는 그룹입니디."));
+    public List<CharacterThumbnailDtoRes> getCharacters(String groupUUID, String userUUID, String workspaceUUID) {
+        GroupEntity group = groupRepository.findByGroupUUID(groupUUID);
+        checkGroupException(group);
+
+        List<CharacterEntity> characters = characterRepository.findAllByGroupUUIDAndDeletedIsFalse(groupUUID);
 
         List<CharacterThumbnailDtoRes> characterThumbnailInfoList = new ArrayList<>();
 
@@ -98,7 +98,7 @@ public class CharacterServiceImpl implements CharacterService {
     @Transactional(readOnly = true)
     @Override
     public List<CharacterThumbnailDtoRes> getTopCharacter(String workspaceUUID, String userUUID) {
-        List<CharacterEntity> characters = characterRepository.findAllByWorkspaceUUIDAndGroupUUIDIsNull(workspaceUUID);
+        List<CharacterEntity> characters = characterRepository.findAllByWorkspaceUUIDAndGroupUUIDIsNullAndDeletedIsFalse(workspaceUUID);
         List<CharacterThumbnailDtoRes> dto = new ArrayList<>();
 
         for (CharacterEntity character : characters) {
@@ -117,7 +117,7 @@ public class CharacterServiceImpl implements CharacterService {
     @Transactional(readOnly = true)
     @Override
     public List<CharacterSimpleDtoRes> getNoGroupCharacters(String workspaceUUID, String userUUID) {
-        List<CharacterEntity> characters = characterRepository.findAllByWorkspaceUUIDAndGroupUUIDIsNull(workspaceUUID);
+        List<CharacterEntity> characters = characterRepository.findAllByWorkspaceUUIDAndGroupUUIDIsNullAndDeletedIsFalse(workspaceUUID);
         List<CharacterSimpleDtoRes> dto = new ArrayList<>();
 
         for (CharacterEntity character : characters) {
@@ -134,16 +134,22 @@ public class CharacterServiceImpl implements CharacterService {
     @Transactional
     @Override
     public void createCharacter(CharacterCreateDtoReq dto, String userUUID) {
-        String characterUUID = UUID.randomUUID().toString();
+        // UUID 프론트에서 받아오는 값으로 대체
+//        String characterUUID = UUID.randomUUID().toString();
 
         String groupUUID = dto.getGroupUUID();
-        // 그룹이 조회되지 않을 때 그룹UUID를 null로 변경
-        if (groupRepository.findByGroupUUID(groupUUID) == null) {
-            groupUUID = null;
+        GroupEntity group = groupRepository.findByGroupUUID(groupUUID);
+        if (groupUUID != null) {
+            checkGroupException(group);
         }
 
+        // 그룹이 조회되지 않을 때 그룹UUID를 null로 변경
+//        if (groupRepository.findByGroupUUID(groupUUID) == null) {
+//            groupUUID = null;
+//        }
+
         RelationEntity newRelation = RelationEntity.builder()
-            .characterUUID(characterUUID)
+            .characterUUID(dto.getCharacterUUID())
             .characterName(dto.getCharacterName())
             .relations(dto.getRelations())
             .build();
@@ -154,10 +160,9 @@ public class CharacterServiceImpl implements CharacterService {
             .userUUID(userUUID)
             .workspaceUUID(dto.getWorkspaceUUID())
             .groupUUID(groupUUID)
-            .characterUUID(characterUUID)
+            .characterUUID(dto.getCharacterUUID())
             .characterName(dto.getCharacterName())
             .description(dto.getDescription())
-
             .relationship(newRelation)
             .characterImage(dto.getCharacterImage());
 
@@ -171,18 +176,18 @@ public class CharacterServiceImpl implements CharacterService {
 
         // 소속 그룹의 엔티티에 추가
         if (groupUUID != null) {
-            GroupEntity parentGroup = groupRepository.findByGroupUUID(groupUUID);
-            CharacterEntity character = characterRepository.findByCharacterUUID(characterUUID);
-            parentGroup.addChildCharacter(character);
+            CharacterEntity character = characterRepository.findByCharacterUUID(dto.getCharacterUUID());
+            group.addChildCharacter(character);
 
-            groupRepository.save(parentGroup);
+            groupRepository.save(group);
         }
     }
 
     @Transactional
     @Override
-    public void updateCharacter(String characterUUID, CharacterUpdateDtoReq dto, String userUUID) {
+    public void updateCharacter(String characterUUID, CharacterUpdateDtoReq dto, String userUUID, String workspaceUUID) {
         CharacterEntity character = characterRepository.findByCharacterUUID(characterUUID);
+        checkCharacterException(character);
         RelationEntity relation = relationRepository.findByCharacterUUID(characterUUID);
         RelationEntity newRelation;
         CharacterEntityBuilder newCharacter;
@@ -220,10 +225,15 @@ public class CharacterServiceImpl implements CharacterService {
 
     @Transactional
     @Override
-    public void moveCharacter(String characterUUID, String groupUUID, String userUUID) {
+    public void moveCharacter(String characterUUID, String groupUUID, String userUUID, String workspaceUUID) {
         CharacterEntity character = characterRepository.findByCharacterUUID(characterUUID);
+        checkCharacterException(character);
+
         GroupEntity parentGroup = groupRepository.findByGroupUUID(character.getGroupUUID());
+
         GroupEntity newParentGroup = groupRepository.findByGroupUUID(groupUUID);
+        checkGroupException(newParentGroup);
+
         parentGroup.removeChildCharacter(character);
         newParentGroup.addChildCharacter(character);
         character.moveCharacter(groupUUID);
@@ -236,24 +246,33 @@ public class CharacterServiceImpl implements CharacterService {
 
     @Transactional
     @Override
-    public void deleteCharacter(String characterUUID, String userUUID) {
+    public void deleteCharacter(String characterUUID, String userUUID, String workspaceUUID) {
         CharacterEntity character = characterRepository.findByCharacterUUID(characterUUID);
-        GroupEntity parentGroup = groupRepository.findByGroupUUID(character.getGroupUUID());
+
+        if (character == null) {
+            throw new NoSuchElementFoundException("유효하지 않은 캐릭터 UUID 입니다.");
+        }
+        if (character.isDeleted()) {
+            throw new DeletedElementException("이미 삭제된 캐릭터 입니다.");
+        }
 
         character.deleteCharacter();
-        parentGroup.addChildCharacter(character);
-
         characterRepository.save(character);
         relationRepository.deleteByCharacterUUID(characterUUID);
-        groupRepository.save(parentGroup);
+
+        GroupEntity parentGroup = groupRepository.findByGroupUUID(character.getGroupUUID());
+        if (parentGroup != null) {
+            parentGroup.removeChildCharacter(character);
+            groupRepository.save(parentGroup);
+        }
     }
 
     @Transactional
     @Override
     public List<CharacterSearchInfoResDTO> searchCharacter(String workspaceUUID, String characterName) {
         List<CharacterEntity> characters = characterRepository
-                .findAllByWorkspaceUUIDAndCharacterNameLike(workspaceUUID, characterName);
-        System.out.println(characters.size());
+                .findAllByWorkspaceUUIDAndDeletedIsFalseAndCharacterNameLike(workspaceUUID, characterName);
+//        System.out.println(characters.size());
         List<CharacterSearchInfoResDTO> characterInfoList = new ArrayList<>();
 
         for (CharacterEntity character : characters) {
@@ -290,7 +309,7 @@ public class CharacterServiceImpl implements CharacterService {
 
     @Transactional(readOnly = true)
     @Override
-    public List<RelationDtoRes> getRelationships() {
+    public List<RelationDtoRes> getRelationships(String userUUID, String workspaceUUID) {
         List<RelationEntity> allRelation = relationRepository.findAll();
         List<RelationDtoRes> allDto = new ArrayList<>();
 
@@ -300,6 +319,10 @@ public class CharacterServiceImpl implements CharacterService {
             String characterUUID = relation.getCharacterUUID();
             String characterName = relation.getCharacterName();
             CharacterEntity character = characterRepository.findByCharacterUUID(characterUUID);
+            if (character == null) {
+                continue;
+            }
+
             String characterImage = character.getCharacterImage();
             Map<String, Double> characterNode = null;
             if (character.getCharacterNode() != null) {
@@ -391,8 +414,9 @@ public class CharacterServiceImpl implements CharacterService {
 
     @Transactional
     @Override
-    public void moveCharacterNode(String characterUUID, Double x, Double y, String userUUID) {
+    public void moveCharacterNode(String characterUUID, Double x, Double y, String userUUID, String workspaceUUID) {
         CharacterEntity character = characterRepository.findByCharacterUUID(characterUUID);
+        checkCharacterException(character);
 
         if (character.getCharacterNode() == null) {
             Map<String, Double> characterNode = new HashMap<>();
@@ -405,5 +429,20 @@ public class CharacterServiceImpl implements CharacterService {
         characterRepository.save(character);
     }
 
-
+    public void checkCharacterException(CharacterEntity character) {
+        if (character == null) {
+            throw new NoSuchElementFoundException("유효하지 않은 캐릭터 UUID 입니다.");
+        }
+        if (character.isDeleted()) {
+            throw new DeletedElementException("삭제된 캐릭터 입니다.");
+        }
+    }
+    public void checkGroupException(GroupEntity group) {
+        if (group == null) {
+            throw new NoSuchElementFoundException("유효하지 않은 그룹 UUID 입니다.");
+        }
+        if (group.isDeleted()) {
+            throw new DeletedElementException("삭제된 그룹 입니다.");
+        }
+    }
 }
